New implementation of Git

+++ test/git-mirage/test.ml
@@ -1,5 +1,6 @@
 (*
  * Copyright (c) 2013-2017 Thomas Gazagnaire <thomas@gazagnaire.org>
+ * and Romain Calascibetta <romain.calascibetta@gmail.com>
  *
  * Permission to use, copy, modify, and distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -15,11 +16,10 @@
  *)
 
 open Lwt.Infix
-open Test_common
-open Git_mirage
-open Result
 
-(* FIXME: this should probably go somewhere else ... *)
+let () = Printexc.record_backtrace true
+
+open Test_common
 
 let protect_unix_exn = function
   | Unix.Unix_error _ as e -> Lwt.fail (Failure (Printexc.to_string e))
@@ -35,26 +35,26 @@ let safe f x = Lwt.catch (fun () -> f x) ignore_enoent
 let mkdir dirname =
   let rec aux dir =
     if Sys.file_exists dir && Sys.is_directory dir then Lwt.return_unit
-    else (
+    else
       let clear =
-        if Sys.file_exists dir then (
-          safe Lwt_unix.unlink dir
-        ) else
-          Lwt.return_unit
+        if Sys.file_exists dir
+        then safe Lwt_unix.unlink dir
+        else Lwt.return ()
       in
       clear >>= fun () ->
       aux (Filename.dirname dir) >>= fun () ->
       protect (Lwt_unix.mkdir dir) 0o755;
-    ) in
+    in
   aux dirname
 
 let command fmt =
-  Printf.ksprintf (fun str ->
-      Printf.printf "[exec] %s\n%!" str;
+  Printf.ksprintf
+    (fun str ->
+      Fmt.(pf stdout) "[exec] %s\n%!" str;
       let i = Sys.command str in
-      if i <> 0 then Printf.printf "[exec] error %d\n%!" i;
-      ()
-    ) fmt
+      if i <> 0 then Fmt.(pf stderr) "[exec] error %d\n%!" i;
+      ())
+    fmt
 
 let rmdir dir =
   if Sys.os_type = "Win32" then
@@ -63,40 +63,53 @@ let rmdir dir =
     command "rm -rf %S" dir
 
 module M = struct
+  let root = "test-git-mirage-store"
 
-  let root = "test-db"
-
-  let (>>|) x f =
+  let ( >>| ) x f =
     x >>= function
     | Ok x    -> f x
     | Error e -> Fmt.kstrf Lwt.fail_with "%a" FS_unix.pp_write_error e
 
   let init () =
-    rmdir root;
-    mkdir root
+    if Sys.file_exists root then rmdir root;
+    mkdir root >>= fun () -> mkdir Filename.(concat root "temp")
 
   include FS_unix
 
-  let connect () = FS_unix.connect root
+  let path p =
+    if Sys.os_type <> "Win32" then p
+    else
+      let segs = Fpath.(segs (normalize (v p))) in
+      String.concat "/" segs
 
+  let read x p = read x (path p)
+  let size x p = size x (path p)
+  let create x p = create x (path p)
+  let mkdir x p = mkdir x (path p)
+  let destroy x p = destroy x (path p)
+  let stat x p = stat x (path p)
+  let listdir x p = listdir x (path p)
+  let write x p = write x (path p)
+  let connect () = FS_unix.connect root
 end
 
-module S = FS(M)(SHA1_slow)(Git.Inflate.M)
+module Gamma = struct
+  type path = Fpath.t
 
-let suite =
-  {
-    name  = "MIRAGE";
-    init  = M.init;
-    clean = unit;
-    store = (module S);
-    shell = true;
-  }
+  let temp = Fpath.(v M.root / "temp")
+  let current = Fpath.v M.root
+end
 
-let extra = [
-  "SHA1-mirage", Test_store.array (module Git_mirage.SHA1_slow);
-]
+module Fs = Git_mirage.FS.Make(Gamma)(M)
+module MirageStore = Git_mirage.Make(Git_mirage.Lock)(Fs)
+
+let mirage_backend =
+  { name  = "Mirage"
+  ; store = (module MirageStore)
+  ; shell = true }
 
 let () =
-  Test_store.run ~extra "git-mirage" [
-    `Quick, suite;
-  ]
+  let () = Lwt_main.run (M.init ()) in
+  let () = verbose () in
+  Alcotest.run "git-mirage"
+    [ Test_store.suite (`Quick, mirage_backend) ]
